# Conflicts:
#	core/src/main/cfml/context/admin/resources/language/en.xml

+++ core/src/main/cfml/context/admin/ext.functions.cfm
@@ -47,7 +47,6 @@
 
 	<cffunction name="loadCFC" returntype="struct" output="yes">
 		<cfargument name="provider" required="yes" type="string">
-		<cfset systemOutput("deprecated function call:<print-stack-trace>",true,true)>
 		<cfreturn createObject('component',"ExtensionProviderProxy").init(arguments.provider)>
 	</cffunction>
 
@@ -61,7 +60,7 @@
 		<cfloop query="providers">
 			<cfif hash(providers.url) EQ arguments.hashProvider>
 				<cfset detail.provider=loadCFC(providers.url)>
-				<Cfdump var=#detail.provider#>
+				<cfdump var=#detail.provider#>
 				<cfabort>
 				<cfset var apps=detail.provider.listApplications()>
 				<cfset detail.info=detail.provider.getInfo()>
@@ -232,7 +231,7 @@
 			</cfif>				
 
 			<cfcatch>
-				<cfset systemOutput(cfcatch,1,1)>
+				<cflog text="Error parsing extension logo, #cfcatch.message#, [#arguments.src#]" type="error">
 				<cfset local.b64=local.empty>
 			</cfcatch>
 		</cftry>
@@ -275,12 +274,12 @@
 			    returnVariable="local.locals" ;
 			// add column if necessary
 			loop list="#locals.columnlist()#" item="local.k" {
-                if(!qry.columnExists(k)) qry.addColumn(k,[]);
+                if(!qry.columnExists(k)) queryAddColumn(qry,k,[]);
             }
-            qry.addColumn('otherVersions',[]);
+           queryAddColumn(qry,'otherVersions',[]);
 
 			loop query="#locals#" {
-				var row=qry.addrow();
+				var row=queryAddRow(qry);
 				qry.setCell("provider","local",row);
 				loop list="#locals.columnlist()#" item="local.k" {
             		qry.setCell(k,locals[k],row);
@@ -309,7 +308,7 @@
 					arrayAppend(ov,older);
 					qry.otherVersions[row]=ov;
 				}
-				qry.deleteRow(row+1);
+				queryDeleteRow(qry,row+1);
 			}
 
 
@@ -332,18 +331,19 @@
 				continue;
 			}
 
-			// rename older to otherVersions
+			lock name="admin-extension-populate-cols" {
+				// rename older to otherVersions
+				if(queryColumnExists(data.extensions,"older") && !queryColumnExists(data.extensions,"otherVersions")) {
+					data.extensions.addColumn("otherVersions",data.extensions.columnData('older'));
+					data.extensions.deleteColumn("older");
+					//QuerySetColumn(data.extensions,"older","otherVersions");
+				}
 
-			if(queryColumnExists(data.extensions,"older") || !queryColumnExists(data.extensions,"otherVersions")) {
-				data.extensions.addColumn("otherVersions",data.extensions.columnData('older'));
-				data.extensions.deleteColumn("older");
-				//QuerySetColumn(data.extensions,"older","otherVersions");
+				// add missing columns
+				loop list="#data.extensions.columnlist()#" item="local.k" {
+					if(!qry.ColumnExists(k)) qry.addColumn(k,[]);
+				}
 			}
-
-			// add missing columns
-			loop list="#data.extensions.columnlist()#" item="local.k" {
-                if(!qry.ColumnExists(k)) qry.addColumn(k,[]);
-            }
 
 			// add Extensions data
 			var row=0;
@@ -410,7 +410,7 @@
 
             	}
 				else {
-					row=qry.addRow();
+					row=queryAddRow(qry);
 					qry.setCell("provider",provider,row);
 					qry.setCell("lastModified",data.lastModified,row);
 	            	loop list="#data.extensions.columnlist()#" item="local.k" {
@@ -645,6 +645,7 @@
 					&"."&repeatString("0",3-len(sct.minor))&sct.minor
 					&"."&repeatString("0",3-len(sct.micro))&sct.micro
 					&"."&repeatString("0",4-len(sct.qualifier))&sct.qualifier
+					& #sct.keyExists("qualifier_appendix1") && isNumeric(sct.qualifier_appendix1)? "."&repeatString("0",4-len(sct.qualifier_appendix1))&sct.qualifier_appendix1  : ""#
 					&"."&repeatString("0",3-len(sct.qualifier_appendix_nbr))&sct.qualifier_appendix_nbr;
 
 		return sct;
